Parse Quality/Performance as double instead of float
std::stof -> std::stod at the four sites parsing Quality (CLI arg and
BC7's string-SetParameter overload), Performance, and ASTC's Quality.
Makes each consistent with the double-typed member each value is
ultimately stored in.

Doesn't independently fix boundary precision loss on the BC7 CLI
-Quality path — compress.cpp casts the parsed value to CODECFLOAT
(typedef float) before it reaches the codec, which reintroduces the
same truncation regardless of parse precision. That's handled on the
bc7enc_rdo adapter side via integer-centibucket comparison (previous
commit). Kept here as a small, independently-correct cleanup —
self-consistent typing, and closes the precision loss on paths that
don't go through the CODECFLOAT cast (e.g. -Performance, which routes
through the string-SetParameter overload directly).

+++ applications/_plugins/common/cmdline.cpp
@@ -408,7 +408,7 @@ static bool ProcessCMDLineOptions(const char* strCommand, const char* strParamet
             {
                 throw "No Quality value specified";
             }
-            float value = std::stof(strParameter);
+            double value = std::stod(strParameter);
             if ((value < 0) || (value > 1.0))
             {
                 throw "Quality value should be in range of 0 to 1.0";
