+++ cmp_compressonatorlib/bc7/codec_bc7.cpp
@@ -128,7 +128,7 @@ bool CCodec_BC7::SetParameter(const CMP_CHAR* pszParamName, CMP_CHAR* sValue)
     }
     if (strcmp(pszParamName, "Quality") == 0)
     {
-        m_Quality = std::stof(sValue);
+        m_Quality = std::stod(sValue);
         if ((m_Quality < 0) || (m_Quality > 1.0))
         {
             return false;
@@ -136,7 +136,7 @@ bool CCodec_BC7::SetParameter(const CMP_CHAR* pszParamName, CMP_CHAR* sValue)
     }
     else if (strcmp(pszParamName, "Performance") == 0)
     {
-        m_Performance = std::stof(sValue);
+        m_Performance = std::stod(sValue);
         if ((m_Performance < 0) || (m_Performance > 1.0))
         {
             return false;
